@@ -27,8 +27,9 @@
 
 package com.coppel.mx.infraestructure.exception;
 
-import com.coppel.mx.model.dto.ErrorMessageDefaultDto;
-
+import com.coppel.mx.model.dto.output.ErrorMessageDefaultDto;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.web.HttpRequestMethodNotSupportedException;
@@ -43,11 +44,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
 
-import javax.servlet.http.HttpServletRequest;
-
 import java.nio.file.AccessDeniedException;
-import java.util.HashMap;
-import java.util.Map;
 
 /**
  * The class Controller exception handler.
@@ -55,93 +52,150 @@ import java.util.Map;
 @ControllerAdvice
 public class ExceptionHandlerController {
 
-  /**
-   * Not found request error exception.
-   *
-   * @param request   the request
-   * @param exception the exception generated
-   * @return The ErrorMessageDefault dto with the error details
-   */
-  @ResponseStatus(HttpStatus.NOT_FOUND)
-  @ExceptionHandler({HttpClientErrorException.class})
-  @ResponseBody
-  public ErrorMessageDefaultDto notFoundRequest(HttpServletRequest request, Exception exception) {
-    return new ErrorMessageDefaultDto(exception, request.getRequestURI());
-  }
+    private static final Logger LOGGER = LoggerFactory.getLogger(ExceptionHandlerController.class);
+    private static final String MESSAGE_FAILED = "FAILED";
 
-  /**
-   * Bad request error exception.
-   *
-   * @param request   the request
-   * @param exception the exception generated
-   * @return The ErrorMessageDefault dto with the error details
-   */
-  @ResponseStatus(HttpStatus.BAD_REQUEST)
-  @ExceptionHandler({
+    /**
+     * Not found request error exception.
+     *
+     * @param exception the exception generated
+     * @return The ErrorMessageDefault dto with the error details
+     */
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    @ExceptionHandler({HttpClientErrorException.class})
+    @ResponseBody
+    public ErrorMessageDefaultDto notFoundRequest(Exception exception) {
+        LOGGER.info(exception.getMessage());
+        return new ErrorMessageDefaultDto(
+                MESSAGE_FAILED,
+                String.valueOf(HttpStatus.NOT_FOUND.value()),
+                HttpStatus.NOT_FOUND.getReasonPhrase());
+    }
+
+    /**
+     * Bad request error exception.
+     *
+     * @param exception the exception generated
+     * @return The ErrorMessageDefault dto with the error details
+     */
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    @ExceptionHandler({
             HttpRequestMethodNotSupportedException.class,
             MissingPathVariableException.class,
             MissingRequestHeaderException.class,
             MissingServletRequestParameterException.class,
             MethodArgumentTypeMismatchException.class,
             HttpMessageNotReadableException.class})
-  @ResponseBody
-  public ErrorMessageDefaultDto badRequest(HttpServletRequest request, Exception exception) {
-    return new ErrorMessageDefaultDto(exception, request.getRequestURI());
-  }
+    @ResponseBody
+    public ErrorMessageDefaultDto badRequest(Exception exception) {
+        LOGGER.info(exception.getMessage());
+        return new ErrorMessageDefaultDto(
+                MESSAGE_FAILED,
+                String.valueOf(HttpStatus.BAD_REQUEST.value()),
+                HttpStatus.BAD_REQUEST.getReasonPhrase());
+    }
 
-  /**
-   * Bad request error validation exception.
-   *
-   * @param exception the exception generated
-   * @return The ErrorMessageDefault dto with the error details
-   */
-  @ResponseStatus(HttpStatus.BAD_REQUEST)
-  @ExceptionHandler(MethodArgumentNotValidException.class)
-  @ResponseBody
-  public ErrorMessageDefaultDto badRequestValid(HttpServletRequest request, MethodArgumentNotValidException exception) {
-    Map<String, String> mapErros = new HashMap<>();
-    exception.getBindingResult().getFieldErrors().forEach((error) -> {
-      String fielName = error.getField();
-      String message = error.getDefaultMessage();
-      mapErros.put(fielName, message);
-    });
-    return new ErrorMessageDefaultDto(exception, request.getRequestURI(), mapErros);
-  }
+    /**
+     * Bad request error validation exception.
+     *
+     * @param exception the exception generated
+     * @return The ErrorMessageDefault dto with the error details
+     */
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    @ResponseBody
+    public ErrorMessageDefaultDto badRequestValid(MethodArgumentNotValidException exception) {
+        LOGGER.info(exception.getMessage());
+        StringBuilder stringBuilder = new StringBuilder();
+        exception.getBindingResult().getFieldErrors().forEach(error -> {
+            String fielName = error.getField();
+            String message = error.getDefaultMessage();
+            stringBuilder.append("field " + fielName + ": error -> " + message + ", ");
+        });
+        return new ErrorMessageDefaultDto(
+                MESSAGE_FAILED,
+                String.valueOf(HttpStatus.BAD_REQUEST.value()),
+                stringBuilder.toString());
+    }
+
+    /**
+     * Unautorized error exception.
+     *
+     * @param exception the exception
+     * @return the error message default dto
+     */
+    //TODO: agregar despues las excepciones de spring security
+    /*@ResponseStatus(HttpStatus.UNAUTHORIZED)
+    @ExceptionHandler({
+    })
+    @ResponseBody
+    public ErrorMessageDefaultDto unautorized(MethodArgumentNotValidException exception) {
+        LOGGER.info(exception.getMessage());
+        return new ErrorMessageDefaultDto(
+               MESSAGE_FAILED,
+               String.valueOf(HttpStatus.UNAUTHORIZED.value()),
+               HttpStatus.UNAUTHORIZED.getReasonPhrase());
+    }*/
 
 
-  /**
-   * Forbidden request error exception.
-   *
-   * @param request   the request
-   * @param exception the exception generated
-   * @return The ErrorMessageDefault dto with the error details
-   */
-  @ResponseStatus(HttpStatus.FORBIDDEN)
-  @ExceptionHandler({
+    /**
+     * Unautorized error exception.
+     *
+     * @param exception the exception
+     * @return the error message default dto
+     */
+    @ResponseStatus(HttpStatus.TOO_MANY_REQUESTS)
+    @ExceptionHandler({
+            HttpClientErrorException.TooManyRequests.class
+    })
+    @ResponseBody
+    public ErrorMessageDefaultDto unautorized(MethodArgumentNotValidException exception) {
+        LOGGER.info(exception.getMessage());
+        return new ErrorMessageDefaultDto(
+                MESSAGE_FAILED,
+                String.valueOf(HttpStatus.TOO_MANY_REQUESTS.value()),
+                HttpStatus.TOO_MANY_REQUESTS.getReasonPhrase());
+    }
+
+
+    /**
+     * Forbidden request error exception.
+     *
+     * @return The ErrorMessageDefault dto with the error details
+     */
+    @ResponseStatus(HttpStatus.FORBIDDEN)
+    @ExceptionHandler({
             AccessDeniedException.class
-  })
-  @ResponseBody
-  public ErrorMessageDefaultDto forbiddenRequest(HttpServletRequest request, Exception exception) {
-    return new ErrorMessageDefaultDto(exception, request.getRequestURI());
-  }
+    })
+    @ResponseBody
+    public ErrorMessageDefaultDto forbiddenRequest(Exception exception) {
+        LOGGER.info(exception.getMessage());
+        return new ErrorMessageDefaultDto(
+                MESSAGE_FAILED,
+                String.valueOf(HttpStatus.FORBIDDEN.value()),
+                HttpStatus.FORBIDDEN.getReasonPhrase());
+    }
 
 
-  /**
-   * Fatal error error exception
-   *
-   * @param request   the request
-   * @param exception the exception generated
-   * @return The ErrorMessageDefault dto with the error details
-   */
-  @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
-  @ExceptionHandler({
+    /**
+     * Fatal error error exception
+     *
+     * @param exception the exception generated
+     * @return The ErrorMessageDefault dto with the error details
+     */
+    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+    @ExceptionHandler({
             Exception.class,
             BusinessRuleException.class,
             InternalServerException.class
-  })
-  @ResponseBody
-  public ErrorMessageDefaultDto fatalError(HttpServletRequest request, Exception exception) {
-    return new ErrorMessageDefaultDto(exception, request.getRequestURI());
-  }
+    })
+    @ResponseBody
+    public ErrorMessageDefaultDto fatalError(Exception exception) {
+        LOGGER.info(exception.getMessage());
+        return new ErrorMessageDefaultDto(
+                MESSAGE_FAILED,
+                String.valueOf(HttpStatus.INTERNAL_SERVER_ERROR.value()),
+                HttpStatus.INTERNAL_SERVER_ERROR.getReasonPhrase());
+    }
 
 }
