@@ -20,45 +20,34 @@
  * any other work released this way by its authors.  You can apply it to
  * your programs, too.
  *
- * Nombre de archivo: HealthController.java
- * Autor: salvgonz
- * Fecha de creación: Mar 13, 2021
+ * Nombre de archivo: CountryDto.java
+ * Autor: gumaro
+ * Fecha de creación: Ago 17, 2023
  */
 
-package com.coppel.mx.application.controller;
 
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
+package com.coppel.mx.model.dto.input;
 
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.validation.constraints.NotNull;
 
 /**
- * The Class HealthController.
+ * The class Country dto.
  */
-@Api(
-        value = "Health API",
-        description = "A service that gives information about service status",
-        tags = {"Health"})
-@RestController
-public class HealthController {
+@Getter
+@Setter
+public class CountryDto {
 
   /**
-   * The Constant MESSAGE.
+   * The Code.
    */
-  private static final  String MESSAGE = "SERVICE UP.";
-
+  @NotNull
+  private String code;
   /**
-   * Health.
-   *
-   * @return the response entity
+   * The Name.
    */
-  @ApiOperation(value = "Check health")
-  @GetMapping(value = "/health")
-  public ResponseEntity<String> health() {
-    return new ResponseEntity<>(MESSAGE, HttpStatus.OK);
-  }
-
+  @NotNull
+  private String name;
 }
