@@ -20,45 +20,26 @@
  * any other work released this way by its authors.  You can apply it to
  * your programs, too.
  *
- * Nombre de archivo: HealthController.java
- * Autor: salvgonz
- * Fecha de creación: Mar 13, 2021
+ * Nombre de archivo: ResponseEvaluateContactDto.java
+ * Autor: gumaro
+ * Fecha de creación: Ago 17, 2023
  */
 
-package com.coppel.mx.application.controller;
+package com.coppel.mx.model.dto.output;
 
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
+import com.coppel.mx.model.dto.input.ContactPointDto;
+import lombok.Getter;
+import lombok.Setter;
 
 /**
- * The Class HealthController.
+ * The class Response evaluate contact dto.
  */
-@Api(
-        value = "Health API",
-        description = "A service that gives information about service status",
-        tags = {"Health"})
-@RestController
-public class HealthController {
+@Getter
+@Setter
+public class ResponseEvaluateContactDto {
 
   /**
-   * The Constant MESSAGE.
+   * The Contact point.
    */
-  private static final  String MESSAGE = "SERVICE UP.";
-
-  /**
-   * Health.
-   *
-   * @return the response entity
-   */
-  @ApiOperation(value = "Check health")
-  @GetMapping(value = "/health")
-  public ResponseEntity<String> health() {
-    return new ResponseEntity<>(MESSAGE, HttpStatus.OK);
-  }
-
+  private ContactPointDto contactPoint;
 }
