@@ -20,45 +20,53 @@
  * any other work released this way by its authors.  You can apply it to
  * your programs, too.
  *
- * Nombre de archivo: HealthController.java
- * Autor: salvgonz
- * Fecha de creación: Mar 13, 2021
+ * Nombre de archivo: SiTelefonosActual.java
+ * Autor: gumaro
+ * Fecha de creación: Ago 17, 2023
  */
 
-package com.coppel.mx.application.controller;
+package com.coppel.mx.model.entity;
 
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
+import lombok.Getter;
+import lombok.Setter;
 
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.Table;
+import java.math.BigInteger;
 
 /**
- * The Class HealthController.
+ * The class Si telefonos actual.
  */
-@Api(
-        value = "Health API",
-        description = "A service that gives information about service status",
-        tags = {"Health"})
-@RestController
-public class HealthController {
+@Entity
+@Getter
+@Setter
+@Table(name = "si_telefonos_actual")
+public class SiTelefonosActual {
 
   /**
-   * The Constant MESSAGE.
+   * The Id.
    */
-  private static final  String MESSAGE = "SERVICE UP.";
+  @Id
+  @Column(name = "id")
+  private BigInteger id;
 
   /**
-   * Health.
-   *
-   * @return the response entity
+   * The Tipo tel.
    */
-  @ApiOperation(value = "Check health")
-  @GetMapping(value = "/health")
-  public ResponseEntity<String> health() {
-    return new ResponseEntity<>(MESSAGE, HttpStatus.OK);
-  }
+  @Column(name = "tipo_tel")
+  private String tipoTel;
+  /**
+   * The Telefono.
+   */
+  @Column(name = "telefono")
+  private String telefono;
+
+  /**
+   * The Estatus tel.
+   */
+  @Column(name = "status_tel")
+  private String estatusTel;
 
 }
